+++ client/src/views/main/main.ts
@@ -1,9 +1,17 @@
-import type { Observer } from '@/lib/observer';
+import type { Observer } from '@/core/ui/observer';
 
 export default class Main implements Observer {
   template(state: StoreState): string {
     return `
-      <div>메인페이지 ${state.router.pathname}</div>
+      <div class="main-page">
+        <div>
+          <label for="date">
+            일자
+            <input name="date">20210720</input>
+          </label>
+        </div>
+        <div></div>
+      </div>
     `;
   }
